modify: add goalChart file and move related files
