+++ src/common/components/dietary-record/goalChart/GoalCompletionRate.tsx
